[EDIT]

+++ rpg/src/actions/action_menu_page.c
@@ -23,8 +23,8 @@ static void action_button_quit(win_t *win)
 static void action_button_help(win_t *win, spt_t *spt, btn_t *btn)
 {
     sfFloatRect limit = sfRectangleShape_getGlobalBounds(btn[10].rect);
-    int mouse_x = win[0].event.mouseButton.x;
-    int mouse_y = win[0].event.mouseButton.y;
+    float mouse_x = win[0].event.mouseButton.x;
+    float mouse_y = win[0].event.mouseButton.y;
 
     if (sfFloatRect_contains(&limit, mouse_x, mouse_y)) {
         help_page(win, spt);
